+++ .prettierignore
@@ -16,5 +16,4 @@ public
 packages/prisma/zod
 packages/prisma/enums
 apps/web/public/embed
-apps/api/v2/swagger/documentation.json
 packages/ui/components/icon/dynamicIconImports.tsx
